+++ test/fixtures/extension-methods-system/src/index.ts
@@ -10,14 +10,13 @@ export function main(): void {
   const s = "hello";
 
   const a = (s as unknown as Ext<string>).asSpan();
-  const b = (s as unknown as Ext<string>).asSpan(1 as int);
+  const b = (s as unknown as Ext<string>).asSpan(1);
 
   const ok1 = a.overlaps(b);
 
-  let off: int = 0 as int;
+  let off: int = 0;
   const ok2 = a.overlaps(b, off);
 
   Console.writeLine(`ok1: ${ok1}`);
   Console.writeLine(`ok2: ${ok2} off: ${off}`);
 }
-
